+++ FacebookAudienceNetwork/com/mopub/mobileads/FacebookInterstitial.java
@@ -8,6 +8,7 @@ import android.text.TextUtils;
 import com.facebook.ads.Ad;
 import com.facebook.ads.AdError;
 import com.facebook.ads.AdSettings;
+import com.facebook.ads.AudienceNetworkAds;
 import com.facebook.ads.InterstitialAd;
 import com.facebook.ads.InterstitialAdListener;
 import com.mopub.common.DataKeys;
@@ -15,6 +16,7 @@ import com.mopub.common.MoPub;
 import com.mopub.common.logging.MoPubLog;
 
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import static com.mopub.mobileads.MoPubErrorCode.EXPIRED;
 
@@ -26,6 +28,7 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
     @NonNull
     private Handler mHandler;
     private Runnable mAdExpiration;
+    private static AtomicBoolean sIsInitialized = new AtomicBoolean(false);
 
     public FacebookInterstitial() {
         mHandler = new Handler();
@@ -53,7 +56,9 @@ public class FacebookInterstitial extends CustomEventInterstitial implements Int
                                     final CustomEventInterstitialListener customEventInterstitialListener,
                                     final Map<String, Object> localExtras,
                                     final Map<String, String> serverExtras) {
-
+        if(!sIsInitialized.getAndSet(true)) {
+            AudienceNetworkAds.initialize(context);
+        }
         setAutomaticImpressionAndClickTracking(false);
 
         MoPubLog.d("Loading Facebook interstitial");